requirements.txt
asn1crypto==0.24.0
cairocffi==0.9.0
CairoSVG==2.2.1
cffi==1.11.5
cryptography==2.4.2
cssselect2==0.2.1
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.3
django-import-export==1.1.0
et-xmlfile==1.0.1
google-images-download==2.5.0
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.10
peewee==3.8.0
Pillow==5.3.0
portalocker==1.3.0
pycparser==2.19
PyMySQL==0.9.2
Pyphen==0.9.5
pypiwin32==223
python-slugify==1.2.6
pytz==2018.7
pywin32==224
PyYAML==3.13
redis==2.10.6
selenium==3.141.0
six==1.11.0
sorl-thumbnail==12.5.0
tablib==0.12.1
tinycss2==0.6.1
unicodecsv==0.14.1
Unidecode==1.0.22
urllib3==1.24.1
webencodings==0.5.1
whitenoise==4.1
xlrd==1.1.0
xlwt==1.3.0
